* add option to include bitmap data into uidesc file 	* add 2 new platform functions to IPlatformBitmap: createFromMemory(..) and createMemoryPNGRepresentation (..) 	* the CFrame object should now be created with the constructor CFrame (size, editor) and afterwards opened with 	  the native parent object and a type description of the parent. The old constructor is now deprecated. 	* add new Bitmap Filter : ScaleLinear

+++ trunk/vstgui/lib/platform/win32/direct2d/d2dfont.cpp
@@ -96,7 +96,7 @@ IDWriteTextLayout* D2DFont::createTextLayout (const CString& string)
 	const WinString* winString = dynamic_cast<const WinString*> (string.getPlatformString ());
 	IDWriteTextLayout* textLayout = 0;
 	if (winString)
-		getDWriteFactory ()->CreateTextLayout (winString->getWideString (), wcslen (winString->getWideString ()), textFormat, 10000, 1000, &textLayout);
+		getDWriteFactory ()->CreateTextLayout (winString->getWideString (), (UINT32)wcslen (winString->getWideString ()), textFormat, 10000, 1000, &textLayout);
 	return textLayout;
 }
